Core: Move the config parser to settings.cpp
