+++ src/core/settings.h
@@ -1,0 +1,6 @@
+#ifndef SETTINGS_H
+#define SETTINGS_H
+
+VSMap *readSettings(const std::string &path);
+
+#endif
